requirements.txt
biopython
numpy
nose
boltons
pytest
